requirements.txt
ipykernel==6.15.0
ipython==8.14.0
ipywidgets==8.1.1
matplotlib==3.7.2
numpy==1.26.2
opencv-python==4.8.0.74
pandas==2.1.4
pillow==10.0.0
pytorch-fid==0.2.0
scipy==1.10.1
tensorboard==2.13.0
torch==1.13.0
torch-summary==1.4.5
torchvision==0.14.0
tqdm==4.65.0